test(cli): split completion stall window
